refactor(core): move dtype from EngineArgs to ModelConfig.from_hf
dtype is a property of the model architecture (the pretrained weights
determine it), not a user-facing engine choice.  ModelConfig.from_hf()
now reads ``torch_dtype`` from the HF config.json when the caller
doesn't explicitly override it.  EngineArgs.dtype removed.

llm_service.py: auto_calculate_num_blocks now uses model_config.dtype
instead of the removed engine_args.dtype.

+++ testbench.py
@@ -168,7 +168,7 @@ def main():
 
     engine_args = EngineArgs(
         model_path="./models/Qwen3-0.6B",
-        device="cuda", block_size=16, dtype=torch.bfloat16,
+        device="cuda", block_size=16,
     )
     llm_service = LLMService.from_engine_args(engine_args)
